수정 중

+++ src/board/board.resolver.ts
@@ -1,10 +1,11 @@
-import { HttpException } from '@nestjs/common';
+import { HttpException, NotFoundException } from '@nestjs/common';
 import { Resolver, Query, Args, Mutation, ID } from '@nestjs/graphql';
 import { UpdateResult } from 'typeorm';
 import { BoardEntity } from './board.entity';
 import { BoardService } from './board.service';
 import { BoardDto } from './dto/board.dto';
 import { BoardInput } from './dto/board.input';
+import { UserInputError } from 'apollo-server-express';
 
 
 @Resolver('Board')
@@ -27,12 +28,19 @@ export class BoardResolver {
     }
 
     @Mutation(() => BoardDto)
-    async update(@Args('id', {type: () => ID}) id: number, @Args('data') data: BoardInput): Promise<void> {
-        await this.boardService.update(id, data)
+    async update(@Args('id') id: number, @Args('data') data: BoardInput): Promise<BoardInput> {
+        return this.boardService.update(id, data)
     }
 
-    @Mutation(() => BoardDto)
-    async delete(@Args('id', {type: () => ID}) id:number): Promise<void>{
-        await this.boardService.delete(id)
+    @Mutation(() => Boolean)
+    async delete(@Args('id') id:number): Promise<Boolean>{
+        console.log(id)
+        if (id) {
+            this.boardService.delete(id)
+            return true
+        } else {
+            throw new NotFoundException('Do Not Exists ID')
+        }
+
     }
 }
