+++ src/board/board.service.ts
@@ -1,6 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { Repository, UpdateResult } from 'typeorm';
+import { Repository } from 'typeorm';
 import { BoardEntity } from './board.entity';
 import { BoardDto } from './dto/board.dto';
 import { BoardInput } from './dto/board.input';
@@ -16,8 +16,7 @@ export class BoardService {
         const board = new BoardEntity();
         board.writer = boardInput.writer;
         board.title = boardInput.title;
-        board.text = boardInput.text;
-        board.date = boardInput.date;       
+        board.text = boardInput.text;      
         return await this.boardRepository.save(board)
     }
 
@@ -31,11 +30,21 @@ export class BoardService {
         return await this.boardRepository.find({ writer: writer });
     }
 
-    async update(id: number, update:BoardInput): Promise<void>{
-       await this.boardRepository.update(id, {...update})
+    async update(id: number, update:BoardInput): Promise<BoardInput>{
+       await this.boardRepository.update(id, update)
     }
 
     async delete(id: number): Promise<void>{
-       await this.boardRepository.delete(id);
+         
+        if (this.boardRepository.findOne(id)) {
+            await this.boardRepository.delete(id)
+        } else {
+            throw new NotFoundException('do not exists')
+        }
+ 
+        
+        
+
+       
     }
 }
